apply astrometry-svn-24610M.patch from Andrew Hood
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@24734 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/an-opts.c
@@ -71,7 +71,7 @@ int opts_getopt(bl* opts, int argc, char** argv) {
         an_option_t* opt = bl_access(opts, i);
         if (!opt->shortopt)
             continue;
-        if (iscntrl(opt->shortopt))
+        if (iscntrl((unsigned)(opt->shortopt)))
 			continue;
         optstring[j] = opt->shortopt;
         j++;
